backend/src/handlers/requirements.txt
boto3>=1.34.0
aws-lambda-powertools>=2.24.0
requests>=2.31.0
Pillow>=10.0.0
ImageHash>=4.3.1